Rider suggested cleanup, part 1  (#5265)

+++ com.unity.ml-agents/Runtime/Inference/BarracudaModelExtensions.cs
@@ -226,8 +226,7 @@ namespace Unity.MLAgents.Inference
             }
             else
             {
-                return model.outputs.Contains(TensorNames.DiscreteActionOutput) &&
-                    (int)model.DiscreteOutputSize() > 0;
+                return model.outputs.Contains(TensorNames.DiscreteActionOutput) && model.DiscreteOutputSize() > 0;
             }
         }
